Pixmap Exporter, added a display of the size in pixels of the image to the exporter dialog
git-svn-id: svn://scribus.net/branches/Version13x/Scribus@5161 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/plugins/pixmapexport/export.h
@@ -53,7 +53,7 @@ public:
 	/*! \brief Dpi of the exported image */
 	int pageDPI;
 	/*! \brief Enlargement of the exported image... 2x 3x etc. */
-	int enlargement;
+	double enlargement;
 	/*! \brief Quality of the image <0; 100> */
 	int quality;
 	/*! \brief A place for stored images */
